Move forward kinematics from kinematics to solvers folder
